requirements.txt
Django>=4.0
djangorestframework
pandas
numpy
scikit-learn
xgboost
tensorflow
joblib
streamlit
gunicorn
python-multipart
whitenoise
catboost
lightgbm